backend/requirements.txt
Flask
SQLAlchemy
Flask-Migrate
openpyxl
numpy
faiss-cpu
